Add/refactor tests
